==> 10 Selenium, Projekte/projektbsp/02 Cleaning.ipynb <==
import pandas as pd

import matplotlib.pyplot as plt
import matplotlib
plt.style.use('fivethirtyeight')
%matplotlib inline

#Looking at data
df = pd.read_csv('readable.csv')

df.info()

#Cleaning Ad Clicks
df['Ad Clicks'] = df['Ad Clicks'].fillna(value=0)
df['Ad Clicks'] = df['Ad Clicks'].astype(int)
#Cleaning Ad Clicks
df['Ad Impressions'] = df['Ad Impressions'].fillna(value=0)
df['Ad Impressions'] = df['Ad Impressions'].astype(int)

#In Ad spend column making 0 out of NaNs
df['Ad Spend'] = df['Ad Spend'].fillna(value=0)
def fillempties(elem):
    elem = str(elem)
    elem = int(elem.replace(' ', '.').replace(';', '').split('.')[0])
    return elem
df['Ad Spend'] = df['Ad Spend'].apply(fillempties)

#Age

df['Age'].value_counts()
def agemin(elem):
    if elem == 'Age: 18 65+':
        elem = elem.replace('Age: ', '').split(' ')[0]
    else:
        elem = elem.replace('Age: ', '').replace('Age: ', '').split('-')[0].strip()
    return int(elem)
df['AgeMin'] = df['Age'].apply(agemin)

def agemax(elem):
    if elem == 'Age: 18 65+':
        elem = elem.replace('Age: ', '').replace('+', '').split(' ')[1]
    else:
        elem = elem.replace('Age: ', '').replace('Age: ', '').replace('+', '').split('-')[1].strip()
    return int(elem)

df['AgeMax'] = df['Age'].apply(agemax)

# Date
df['Date'] = df['Date'].fillna(value='N/A')
def remove(elem):
    elem = elem[:17]
    if elem == 'N/A':
        return '01/01/01 01:01:01'
    else:
        return elem
df['Date'] = df['Date'].apply(remove)
df.index = pd.to_datetime(df['Date'], format='%m/%d/%y %H:%M:%S')

df[df['Date']!= '01/01/01 01:01:01'].resample('W')['Number'].count().plot()

df['Location'] = df['Location'].fillna(value='N/A')
#Cleaning location column
def defsplit(elem):
    elem = elem.split('Interests:')[0]
    elem = elem.split('Excluded Connections:')[0].replace('\\n','')
    elem = elem.replace('Location - Living  In:', '')
    elem = elem.replace('Location: ', '').replace('Location*', '').strip()
    return elem
df['Cleaned location'] = df['Location'].apply(defsplit)

# Interests
df['Interests'] = df['Interests'].fillna(value='N/A')
def remove(elem):
    elem = elem.replace('Interests: ', '')
    if ('St.' in elem or
        'Jr.' in elem):
        elem = elem.replace(' or ', ',').split(',')
    else: 
        elem = elem.replace(' or ', ','.replace('. ', ',')).split(',')
    return elem
df['Interests'] = df['Interests'].apply(remove)
lst = list(df['Interests'])
flat_list = [item for sublist in lst for item in sublist]
fl = []
for elem in flat_list:
    elem = elem.upper().replace(' OR', '').strip()
    fl.append(elem)
df_tags = pd.DataFrame(fl)
df_tags.columns = ['Tags']
df_tags['Tags'].value_counts().count()

#Language
df['Language'].value_counts()

#Location
df['Location'] = df['Location'].fillna(value='N/A')
def defsplit(elem):
    elem = elem.split('Interests:')[0]
    elem = elem.split('Excluded Connections:')[0].replace('\\n','')
    elem = elem.replace('Location - Living  In:', '')
    elem = elem.replace('Location: ', '').replace('Location*', '').strip()
    return elem
df['Location'] = df['Location'].apply(defsplit)

df['Location'].value_counts()

df.to_csv('readable_cleaned.csv')
